Clean auth and No auth if no .env
